requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
slack-bolt==1.18.0
notion-client==2.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-jose==3.3.0
pydantic==2.5.2
openai==1.3.7
httpx==0.24.1 